Naive Integration of parser in the code

+++ Niroj/LoadSheddingReminder/app/src/main/java/com/bishalniroj/loadsheddingreminder/database/LoadSheddingScheduleDbHelper.java
@@ -5,6 +5,7 @@ import android.database.SQLException;
 
 import com.bishalniroj.loadsheddingreminder.Utilities;
 
+import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
@@ -19,12 +20,14 @@ public class LoadSheddingScheduleDbHelper {
     private int mNumberOfAreas = -1;
     private Utilities.AreaSchedulingInfo[] mAreaInfo;
 
-    private LoadSheddingScheduleCompleteInfoTable mScheduleInfoTable;
+    private static LoadSheddingScheduleCompleteInfoTable mScheduleInfoTable;
     private static LoadSheddingScheduleDbHelper mDbHelper = null;
     private static int mReferenceCount = 0;
 
-	private LoadSheddingScheduleDbHelper(Context cCtx) {
-        mScheduleInfoTable = new LoadSheddingScheduleCompleteInfoTable(cCtx);
+	private LoadSheddingScheduleDbHelper(Context cCtx, Boolean writeFlag) {
+        if(mScheduleInfoTable == null) {
+            mScheduleInfoTable = new LoadSheddingScheduleCompleteInfoTable(cCtx);
+        }
         mAreaInfo = new Utilities.AreaSchedulingInfo[Utilities.MAXIMUM_NUMBER_OF_AREA];
         for( int i=0; i<Utilities.MAXIMUM_NUMBER_OF_AREA; i++ ) {
             mAreaInfo[i] = new Utilities.AreaSchedulingInfo();
@@ -37,16 +40,18 @@ public class LoadSheddingScheduleDbHelper {
 
         //For testing
         mScheduleInfoTable.open();
-        mScheduleInfoTable.dropTable();
-        mScheduleInfoTable.createTable();
-        fillDatabase();
+        if(writeFlag) {
+            mScheduleInfoTable.dropTable();
+            mScheduleInfoTable.createTable();
+        }
+        //fillDatabase();
 	}
 
-    public static LoadSheddingScheduleDbHelper GetInstance(Context ctx ) {
+    public static LoadSheddingScheduleDbHelper GetInstance(Context ctx, Boolean writeFlag ) {
         if( mDbHelper == null ) {
             synchronized(LoadSheddingScheduleDbHelper.class) {
                 if( mDbHelper == null ) {
-                    mDbHelper = new LoadSheddingScheduleDbHelper(ctx);
+                    mDbHelper = new LoadSheddingScheduleDbHelper(ctx, writeFlag);
                 }
             }
         }
@@ -95,8 +100,20 @@ public class LoadSheddingScheduleDbHelper {
         }
 	}
 
+    //TODO: Check how the mAreaInfo mWeekInfo mDailySched is prepopulated for this to work
+    /*
     public ArrayList<Utilities.LoadSheddingScheduleData> GetSchedDataForADay( int areaNum, int day ) {
         return mAreaInfo[areaNum].mWeekInfo[day].mDailySched;
+    }
+    */
+
+    public ArrayList<Utilities.LoadSheddingScheduleData> GetSchedDataForADay( int areaNum, int day ) {
+        return mScheduleInfoTable.getScheduleForADayAndArea(areaNum, day);
+    }
+
+
+    public List<Utilities.LoadSheddingCompleteSchedule> GetAllSchedule(  ) {
+        return mScheduleInfoTable.getAllScheduleInfo();
     }
 
     // Think of better way to do it !!!
@@ -122,5 +139,34 @@ public class LoadSheddingScheduleDbHelper {
             }
         }
     }
+
+    /*
+    This is for testing with real data
+     */
+
+    //fill database with real data
+    public void fillDatabaseReal(ArrayList<ArrayList<ArrayList<Utilities.LoadSheddingScheduleData>>>
+                                  sD)
+    {
+        ArrayList<ArrayList<ArrayList<Utilities.LoadSheddingScheduleData>>> allScheduleData =
+      ( ArrayList<ArrayList<ArrayList<Utilities.LoadSheddingScheduleData>>>)sD.clone();
+        for (int i=0;i<allScheduleData.size();i++){
+        ArrayList<ArrayList<Utilities.LoadSheddingScheduleData>> areaScheduleData = allScheduleData.get(i);
+            for (int j=0; j<areaScheduleData.size();j++){
+                ArrayList<Utilities.LoadSheddingScheduleData> dayScheduleData = areaScheduleData.get(j);
+                for (Utilities.LoadSheddingScheduleData indvSchedule:dayScheduleData){
+                    int startHour = indvSchedule.mStartHour;
+                    int startMins = indvSchedule.mStartMins;
+                    int endHour   = indvSchedule.mEndHour;
+                    int endMins   = indvSchedule.mEndMins;
+                    mScheduleInfoTable.insertScheduleInfo(new Utilities.LoadSheddingCompleteSchedule
+                            (i, j, new Utilities.LoadSheddingScheduleData(startHour, startMins,
+                                    endHour, endMins)));
+                }
+            }
+        }
+    }
+
+
 
 }
